seperated js files as specified and added live link to readme
