+++ README.md
@@ -8,6 +8,9 @@ A weather app using the openweather maps api
 
 - Javascript
 
+## Live Link
+https://marvellousubani.github.io/weather-app (When on Master)
+
 ## How to Run the App Locally
 ```
    git clone https://github.com/MarvellousUbani/weather-app
